refactor: 기존 JWT 구현 방식 의존성 삭제
- 먼저 세션 및 쿠키로 개발 후 JWT로 마이그레이션 예정입니다.
- 테스트 완료 했습니다.
