+++ src/main/java/kr/adapterz/community/auth/controller/AuthController.java
@@ -1,4 +1,4 @@
-package kr.adapterz.community.auth;
+package kr.adapterz.community.auth.controller;
 
 import jakarta.validation.Valid;
 import kr.adapterz.community.auth.dto.LoginRequestDto;
@@ -20,14 +20,14 @@ public class AuthController {
     private final AuthService authService;
 
     @PostMapping
-    public ResponseEntity<ApiResponseDto<LoginResponseDto>> login(
+    public ResponseEntity<ApiResponseDto<LoginResponseDto>> createAuth(
             @Valid @RequestBody LoginRequestDto dto) {
 
-        LoginResponseDto loginResponseDto = authService.createAuth(dto);
+        LoginResponseDto loginResponseDto = authService.login(dto);
         ApiResponseDto<LoginResponseDto> responseBody = ApiResponseDto.success(loginResponseDto,
                 "로그인에 성공하였습니다.");
-        return ResponseEntity.ok()
-//                .header(HttpHeaders.SET_COOKIE)
-                .body(responseBody);
+        // 인증 정보(세션) 생성 및 쿠키 생성
+        // 헤더에 쿠키 첨부
+        return ResponseEntity.ok().body(responseBody);
     }
 }
