+++ src/main/java/kr/adapterz/community/auth/dto/LoginResponseDto.java
@@ -1,6 +1,5 @@
 package kr.adapterz.community.auth.dto;
 
-import kr.adapterz.community.security.jwt.JwtDto;
 import kr.adapterz.community.user.entity.User;
 import lombok.Builder;
 
@@ -12,23 +11,14 @@ public record LoginResponseDto(
 
         String nickname,
 
-        String profileImageUrl,
-
-        String tokenType,
-
-        String accessToken,
-
-        Long expiresIn
+        String profileImageUrl
 ) {
-    public static LoginResponseDto of(String email, User user, JwtDto jwtDto) {
+    public static LoginResponseDto of(String email, User user) {
         return LoginResponseDto.builder()
                 .userId(user.getId())
                 .email(email)
                 .nickname(user.getNickname())
                 .profileImageUrl(user.getProfileImageUrl())
-                .tokenType(jwtDto.tokenType())
-                .accessToken(jwtDto.accessToken())
-                .expiresIn(jwtDto.expiresIn())
                 .build();
     }
 }
