+++ src/main/java/kr/adapterz/community/auth/service/AuthService.java
@@ -5,8 +5,8 @@ import kr.adapterz.community.auth.dto.LoginResponseDto;
 
 public interface AuthService {
     // 로그인, 인증 자원 생성
-    LoginResponseDto createAuth(LoginRequestDto dto);
+    LoginResponseDto login(LoginRequestDto dto);
 
     // 로그아웃, 인증 자원 제거
-    void deleteAuth();
+    void logout();
 }
