+++ src/main/java/kr/adapterz/community/auth/service/AuthServiceImpl.java
@@ -6,8 +6,6 @@ import kr.adapterz.community.auth.dto.LoginRequestDto;
 import kr.adapterz.community.auth.dto.LoginResponseDto;
 import kr.adapterz.community.common.exception.UnauthorizedException;
 import kr.adapterz.community.security.Encoder;
-import kr.adapterz.community.security.jwt.JwtDto;
-import kr.adapterz.community.security.jwt.JwtManager;
 import kr.adapterz.community.user.entity.User;
 import kr.adapterz.community.user.entity.UserAuth;
 import kr.adapterz.community.user.service.UserService;
@@ -19,22 +17,26 @@ import org.springframework.stereotype.Service;
 public class AuthServiceImpl implements AuthService {
     private final UserService userService;
     private final Encoder encoder;
-    private final JwtManager jwtManager;
 
+    /**
+     * 이메일, 비밀번호를 통해 인증을 진행하고, 인가를 위한 인증 정보를 생성하여 클라이언트에게 줍니다.
+     *
+     * @param dto
+     * @return
+     */
     @Override
-    public LoginResponseDto createAuth(LoginRequestDto dto) {
+    public LoginResponseDto login(LoginRequestDto dto) {
         UserAuth userAuth = userService.findUserAuthByEmail(dto.email());
         String passwordHash = userAuth.getPasswordHash();
         if (!encoder.checkPassword(dto.password(), passwordHash)) {
             throw new UnauthorizedException(AUTH_FAILURE);
         }
         User user = userService.findById(userAuth.getUser().getId());
-        JwtDto jwtDto = jwtManager.generateToken(user.getId());
-        return LoginResponseDto.of(userAuth.getEmail(), user, jwtDto);
+        return LoginResponseDto.of(userAuth.getEmail(), user);
     }
 
     @Override
-    public void deleteAuth() {
+    public void logout() {
 
     }
 }
