+++ src/main/java/kr/adapterz/community/user/dto/UserResponseDto.java
@@ -1,6 +1,5 @@
 package kr.adapterz.community.user.dto;
 
-import kr.adapterz.community.security.jwt.JwtDto;
 import kr.adapterz.community.user.entity.User;
 import lombok.Builder;
 
@@ -12,23 +11,14 @@ public record UserResponseDto(
 
         String nickname,
 
-        String profileImageUrl,
-
-        String tokenType,
-
-        String accessToken,
-
-        Long expiresIn
+        String profileImageUrl
 ) {
-    public static UserResponseDto of(User user, String email, JwtDto jwtDto) {
+    public static UserResponseDto of(User user, String email) {
         return UserResponseDto.builder()
                 .userId(user.getId())
                 .email(email)
                 .nickname(user.getNickname())
                 .profileImageUrl(user.getProfileImageUrl())
-                .tokenType(jwtDto.tokenType())
-                .accessToken(jwtDto.accessToken())
-                .expiresIn(jwtDto.expiresIn())
                 .build();
     }
 }
