+++ src/test/java/kr/adapterz/community/user/service/UserServiceTest.java
@@ -11,8 +11,6 @@ import static org.mockito.Mockito.verify;
 
 import kr.adapterz.community.common.exception.BadRequestException;
 import kr.adapterz.community.security.Encoder;
-import kr.adapterz.community.security.jwt.JwtDto;
-import kr.adapterz.community.security.jwt.JwtManager;
 import kr.adapterz.community.user.dto.CreateUserRequestDto;
 import kr.adapterz.community.user.dto.UserResponseDto;
 import kr.adapterz.community.user.entity.User;
@@ -42,8 +40,8 @@ class UserServiceTest {
     @Mock
     private Encoder encoder;
 
-    @Mock
-    private JwtManager jwtManager;
+//    @Mock
+//    private JwtManager jwtManager;
 
     @DisplayName("회원가입 성공 테스트")
     @Test
@@ -53,19 +51,12 @@ class UserServiceTest {
                 "testuser", "url");
         User savedUser = User.from(request);
         ReflectionTestUtils.setField(savedUser, "id", 1L); // savedUser 객체의 'id' 필드에 1L 값을 강제로 주입
-        JwtDto jwtDto = JwtDto.builder()
-                .tokenType("Bearer")
-                .accessToken("example")
-                .refreshToken("example")
-                .expiresIn(3600L)
-                .build();
         // Mock 객체의 행동 정의
         // BDD(Behavior Driven Development)를 따라 mockito 기본 API인 when이 아닌 wrapping한 given을 사용합니다.
         given(userAuthRepository.existsByEmail(request.email())).willReturn(false);
         given(userRepository.existsByNickname(request.nickname())).willReturn(false);
         given(encoder.encodePassword(request.password())).willReturn("encodedPassword");
         given(userRepository.save(any(User.class))).willReturn(savedUser);
-        given(jwtManager.generateToken(savedUser.getId())).willReturn(jwtDto);
 
         // when (무엇을 할 때)
         // @InjectMocks으로 인해 Mock 객체가 주입된 UserService::createUser를 실행하게 됩니다.
